chore: break out ci constants into JS file
