+++ .evergreen/generate_evergreen_tasks.js
@@ -2,23 +2,17 @@ const fs = require('fs');
 const yaml = require('js-yaml');
 const { mongoshTasks } = require('./generate_mongosh_tasks');
 
-const MONGODB_VERSIONS = ['latest', 'rapid', '6.0', '5.0', '4.4', '4.2', '4.0', '3.6'];
-const versions = [
-  { codeName: 'erbium', versionNumber: 12 },
-  { codeName: 'fermium', versionNumber: 14 },
-  { codeName: 'gallium', versionNumber: 16 },
-  { codeName: 'hydrogen', versionNumber: 18 }
-];
-const NODE_VERSIONS = versions.map(({ codeName }) => codeName);
-NODE_VERSIONS.sort();
-const LOWEST_LTS = NODE_VERSIONS[0];
-const LATEST_LTS = NODE_VERSIONS[NODE_VERSIONS.length - 1];
-
-const TOPOLOGIES = ['server', 'replica_set', 'sharded_cluster'];
-const AWS_AUTH_VERSIONS = ['latest', '6.0', '5.0', '4.4'];
-const TLS_VERSIONS = ['latest', '6.0', '5.0', '4.4', '4.2'];
-
-const DEFAULT_OS = 'rhel80-large';
+const {
+  MONGODB_VERSIONS,
+  versions,
+  NODE_VERSIONS,
+  LOWEST_LTS,
+  LATEST_LTS,
+  TOPOLOGIES,
+  AWS_AUTH_VERSIONS,
+  TLS_VERSIONS,
+  DEFAULT_OS
+} = require('./ci_matrix_constants');
 
 const OPERATING_SYSTEMS = [
   {
@@ -368,14 +362,13 @@ for (const VERSION of AWS_AUTH_VERSIONS) {
 
 const BUILD_VARIANTS = [];
 
-for (const
-  {
-    name: osName,
-    display_name: osDisplayName,
-    run_on,
-    nodeVersions = NODE_VERSIONS,
-    clientEncryption,
-  } of OPERATING_SYSTEMS) {
+for (const {
+  name: osName,
+  display_name: osDisplayName,
+  run_on,
+  nodeVersions = NODE_VERSIONS,
+  clientEncryption
+} of OPERATING_SYSTEMS) {
   const testedNodeVersions = NODE_VERSIONS.filter(version => nodeVersions.includes(version));
   const os = osName.split('-')[0];
   const tasks = BASE_TASKS.concat(TASKS).filter(task => {
